pyproject.toml
[tool.poetry]
name = "naotool"
version = "0.1.2"
description = "A packge of utils or tools keeps Python sweet."
authors = ["tomorinao-www <1520872279@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "naotool"}]
homepage = "https://github.com/tomorinao-www/naotool"
repository = "https://github.com/tomorinao-www/naotool"
documentation = "https://github.com/tomorinao-www/naotool#README.md"

[tool.poetry.urls]
"Homepage" = "https://github.com/tomorinao-www/naotool"
"Repository" = "https://github.com/tomorinao-www/naotool"
"Documentation" = "https://github.com/tomorinao-www/naotool#README.md"
"Issue Tracker" = "https://github.com/tomorinao-www/naotool/issues"

[tool.poetry.dependencies]
python = "^3.10"
httpx = ">=0.27.2"
pillow = ">=11.0.0"
typeguard = "^4.3.0"
logging = ">=0.4.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
